% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CaseToS4}
\alias{CaseToS4}
\title{Convert a single case row to a list of S4 objects for this case. Each
Filepath will be converted to a separate object.}
\usage{
CaseToS4(case.row)
}
\description{
Convert a single case row to a list of S4 objects for this case. Each
Filepath will be converted to a separate object.
}
