% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cGroupBy}
\alias{cGroupBy}
\title{Group list of S4 flow entries by slot and optionally limit to tube}
\usage{
cGroupBy(input, group_on, tube_nums)
}
\arguments{
\item{input}{List of flow entries.}

\item{group_on}{String for slot to group on.}

\item{tube_nums}{Integer vector for tube numbers to be used.}
}
\description{
Disable tube limiting by passing an empty vector such as vector().
}
