% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{CreateFileInfo}
\alias{CreateFileInfo}
\title{Read and preprocess file structure}
\usage{
CreateFileInfo(path, thread.num, ext = "LMD", remove.duplicates = T,
  material = NULL)
}
\arguments{
\item{path}{Top directory which contains the lower directories serving as different groups.}

\item{ext}{File extension of files to be included. Generally this will be either lmd or fcs (case sensitive)}

\item{remove.duplicates}{Remove duplicate file entries.}

\item{material}{Optional filter on used material of files.}
}
\value{
File matrix with specified set and no duplicates.
}
\description{
Convenient wapper to extract a single set without duplicates from directory file structure.
}
