% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CreateLapply}
\alias{CreateLapply}
\title{Parallelize functions with lapplies}
\usage{
CreateLapply(thread.num = 1, ...)
}
\arguments{
\item{thread.num}{Number of threads specified for operation. A number of 1
will use normal lapply.}
}
\value{
Correctly parallelized lapply function.
}
\description{
Create a new cluster context for function to be parallelized. This creates
considerable overhead, but is easier to debug and maintain correct object
dependencies in complex lapply functions. All functions needing this should
be condidered for more elaborate preformance optimization.
}
