% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_processing.R
\name{FcsSelectMarkers}
\alias{FcsSelectMarkers}
\title{Reduce flowframe to specified set of marker channels.}
\usage{
FcsSelectMarkers(fcs_entry, selected.markers, ...)
}
\arguments{
\item{fcs_entry}{File matrix containing loaded flowframes.}

\item{selection}{Vector of marker names, which will be used for selection.}
}
\value{
Flowframe with filtered flowframes. Flowframes not having all specified markers will be replaced with NA.
}
\description{
Reduce flowframe to specified set of marker channels.
}
