% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{filter_file_info}
\alias{filter_file_info}
\title{Filter file matrix according to parameters}
\usage{
filter_file_info(file_info, ...)
}
\arguments{
\item{file_info}{File Matrix as loaded by get_dir}

\item{...}{named parameters matching names in the file matrix.}
}
\value{
Filtered file matrix.
}
\description{
Filter file matrix according to parameters
}
