% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{filter_flowFrame_majority}
\alias{filter_flowFrame_majority}
\title{Exclude flowframes without common channels.}
\usage{
filter_flowFrame_majority(flow_entries, threshold, cluster)
}
\arguments{
\item{threshold}{Minimum ratio for marker to be common.}

\item{fcs_info}{File matrix with loaded flowframes.}
}
\value{
File matrix with removed channels.
}
\description{
Exclude flowframes without common channels.
}
\examples{
t = filter_flowFrame_majority(files[1:10], 0.8)
t = filter_flowFrame_majority(files, 0.8, cluster)
}
