% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{filter_flowFrame_markers}
\alias{filter_flowFrame_markers}
\title{Reduce flowframes to specified set of marker channels.}
\usage{
filter_flowFrame_markers(fcs_info, markers)
}
\arguments{
\item{fcs_info}{File matrix containing loaded flowframes.}

\item{markers}{Vector of marker names, which will be used for selection.}
}
\value{
Flowframe with filtered flowframes. Flowframes not having all specified markers will be replaced with NA.
}
\description{
Reduce flowframes to specified set of marker channels.
}
