% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{filter_list}
\alias{filter_list}
\title{Filter list of flowEntries on arbitrary slots}
\usage{
filter_list(file_entries, ...)
}
\arguments{
\item{file_entries}{List of flow entries}

\item{...}{named argument corresponding to a fcs slot}
}
\value{
Filtered list of flowEntries
}
\description{
Filter list of flowEntries on arbitrary slots
}
\examples{
filtered = filter_list(files, tube_set=1, group='normal')
}
