% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{FilterChannelMajority}
\alias{FilterChannelMajority}
\title{Return flowframes with common channels and selected channels.}
\usage{
FilterChannelMajority(flow.entries, threshold = 0.9)
}
\arguments{
\item{flow.entries}{File matrix with loaded flowframes.}

\item{threshold}{Minimum ratio for marker to be common.}
}
\value{
List with filtered list and the names of the selected channels.
}
\description{
Return flowframes with common channels and selected channels.
}
