% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_processing.R
\name{FilterCommonChannels}
\alias{FilterCommonChannels}
\title{Exclude flowframes without common channels.}
\usage{
FilterCommonChannels(flow_entries, ...)
}
\arguments{
\item{fcs_info}{File matrix with loaded flowframes.}

\item{threshold}{Minimum ratio for marker to be common.}
}
\value{
File matrix with removed channels.
}
\description{
Exclude flowframes without common channels.
}
\examples{
t = FilterCommonChannels(files[1:10], 0.8)
t = FilterCommonChannels(files, 0.8, cluster)
}
