% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{FilterEntries}
\alias{FilterEntries}
\title{Return boolean vector for a list based on filter}
\usage{
FilterEntries(entry.list, filters, acc)
}
\arguments{
\item{filters}{List of named vectors.}

\item{acc}{Optional parameter to override the function used to access the
compared value.}

\item{file.list}{List of S4 objects implementing the names in the filter.}
}
\value{
boolean vector used for subsetting the list.
}
\description{
Return boolean vector for a list based on filter
}
\examples{
all.files[FilterOnAttr(all.files, list(group='CLL'))]
}
