% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{FilterEntry}
\alias{FilterEntry}
\title{Return bool for entry depending on whether it matches filter}
\usage{
FilterEntry(entry, filters, acc, ...)
}
\arguments{
\item{entry}{Entry type S4 class.}

\item{filters}{List of vectors containing acceptable values.}

\item{acc}{Function to access the value being compared.}
}
\value{
Boolean whether filter is fulfilled.
}
\description{
Return bool for entry depending on whether it matches filter
}
