% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{FilterList}
\alias{FilterList}
\title{Filter list of flowEntries on arbitrary slots}
\usage{
FilterList(file.entries, ...)
}
\arguments{
\item{file.entries}{List of flow entries}

\item{...}{named argument corresponding to a fcs slot}
}
\value{
Filtered list of flowEntries
}
\description{
Filter using argument names instead of filters list
}
\examples{
filtered = FilterList(files, tube_set=1, group='normal')
}
