% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{class}
\name{FlowEntry-class}
\alias{FlowEntry-class}
\alias{FlowEntry}
\title{S4 class representing a flowframe with additional metadata}
\description{
S4 class representing a flowframe with additional metadata
}
\section{Slots}{

\describe{
\item{\code{filepath}}{Path to fcs file.}

\item{\code{group}}{Group of single case. Most often corresponding to the disease label.}

\item{\code{label}}{Unique identifier matching a single case.}

\item{\code{material}}{Sample material used in analysis.}

\item{\code{tube_set}}{Multiple tubes can be used to capture a larger number of channels per sample.}

\item{\code{fcs}}{Slot to load fcs file.}
}}

