% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{get_dir}
\alias{get_dir}
\title{Get file structure information from a given directory, using the lower directories as groupings.}
\usage{
get_dir(path, ext)
}
\arguments{
\item{path}{Top directory which contains the lower directories serving as different groups.}

\item{ext}{File extension of files to be included. Generally this will be either lmd or fcs (case sensitive)}
}
\value{
Matrix containing filepath, group, label (regex currently hardcoded for specific naming schemata)
}
\description{
Get file structure information from a given directory, using the lower directories as groupings.
}
\examples{
get_dir('/data/flowData', 'FCS')
get_dir('../data/fcs', 'lmd')
}
