% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{GetDir}
\alias{GetDir}
\title{Get file structure information from a given directory, using the lower directories as groupings.}
\usage{
GetDir(path, ext = "LMD", ...)
}
\arguments{
\item{path}{Top directory which contains the lower directories serving as different groups.}

\item{ext}{File extension of files to be included. Generally this will be either lmd or fcs (case sensitive)}
}
\value{
Matrix containing filepath, group, label (regex currently hardcoded for specific naming schemata)
}
\description{
Get file structure information from a given directory, using the lower directories as groupings.
}
\examples{
GetDir('/data/flowData', 'FCS')
GetDir('../data/fcs', 'lmd')
files = GetDir(testdir, 'LMD', cluster)
files = GetDir(testdir, 'LMD')
}
