% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{GetFile}
\alias{GetFile}
\title{Automatically check if file needs to be downloaded prior to usage}
\usage{
GetFile(filename, dirpath, ...)
}
\description{
Build a filename and automatically try to resolve remote filepaths
}
