% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{JsonToS4}
\alias{JsonToS4}
\title{Convert list of data from json to S4 objects
Keep grouping of cohort - case-id - s4-objs}
\usage{
JsonToS4(json.list, temp.path = "", dir.path = "")
}
\description{
Convert list of data from json to S4 objects
Keep grouping of cohort - case-id - s4-objs
}
