% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{majority_markers}
\alias{majority_markers}
\title{Determine marker names available in all flow data.}
\usage{
majority_markers(fcs_info, threshold = 0.95)
}
\arguments{
\item{fcs_info}{File matrix with loaded flowFrames.}

\item{threshold}{Minimal ratio of the availiability of one marker to all flowframes to be included.}
}
\value{
Vector with marker names available in a larger than threshold ratio of flowframes.
}
\description{
Determine marker names available in all flow data.
}
\examples{
fcs_data = load_fcs(file_info, threads=8)
majority_markers(fcs_data, threshold=0.8)
}
