% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{marker_occurrences}
\alias{marker_occurrences}
\title{Get distribution of marker channels across whole dataset.}
\usage{
marker_occurrences(flow_entries, cluster)
}
\arguments{
\item{flow_entries}{Flow entry list, FCS do not have to be loaded}
}
\value{
Table with occurrences of marker channels as 1/0 matrix.
}
\description{
Get distribution of marker channels across whole dataset.
}
\examples{
t = marker_occurrences(files, cluster)
t = marker_occurrences(files[1:10])
}
