% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_processing.R
\name{MarkerOccurences}
\alias{MarkerOccurences}
\title{Get distribution of marker channels across whole dataset.}
\usage{
MarkerOccurences(flow.entries)
}
\arguments{
\item{flow_entries}{Flow entry list with loaded fcs files.}
}
\value{
Table with occurrences of marker channels as 1/0 matrix.
}
\description{
Get distribution of marker channels across whole dataset.
}
\examples{
t = marker_occurrences(files, cluster)
t = marker_occurrences(files[1:10])
}
