% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{PathExists}
\alias{PathExists}
\title{Check if path exists}
\usage{
PathExists(path, true.fun = NULL, false.fun = NULL, ...)
}
\description{
Supports remote addresses. Executes given functions and return boolean truth
values.
}
