% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{process_dir}
\alias{process_dir}
\title{Generate file matrix from directory}
\usage{
process_dir(path, ext = "LMD", set = 1, threshold = 0.9,
  group_size = 50, threads = 1, simple_marker_names = FALSE)
}
\arguments{
\item{path}{Top directory which contains the lower directories serving as different groups.}

\item{ext}{File extension of files to be included. Generally this will be either lmd or fcs (case sensitive)}

\item{set}{Specifies set which will be returned.}

\item{threads}{Number of cpu threads used in file reading.}

\item{simple_marker_names}{Use antibody name as marker name.}
}
\value{
File matrix with loaded and filtered datasets.
}
\description{
A file matrix contains information about each file included in the structure
such as group identity, set (used for different experiment tubes for the
flow cytometry information). This structure serves as the basis for later
clustering and dimension reduction steps.
}
\section{File Matrix description}{

filepath - Name of the file concatenated with the provided path
group - Group as specified by the subfolder name
set - Additional descriptor identifying multiple files per label. (denotes tube)
label - ID label from the filename
fcs - flowFrame data from the corresponding file
}

