% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_processing.R
\name{ProcessSingle}
\alias{ProcessSingle}
\title{Load and filter single file info entry}
\usage{
ProcessSingle(file.entry, selected.channels, trans = "logicle",
  remove_margins = TRUE, ...)
}
\arguments{
\item{file.entry}{File info row from a file matrix.}

\item{selected.channels}{Marker names selected from flowframe.}

\item{trans}{String name for transformation function.}

\item{remove_margins}{Boolean for removal of events with 0 or maximum values.}
}
\value{
File row or NULL if marker name not in flowframe.
}
\description{
Load and filter single file info entry
}
\examples{
t = ProcessSingle(files[[1]],selection)
}
