% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{PutFile}
\alias{PutFile}
\title{Automatically check if outputpath needs uploading to S3}
\usage{
PutFile(fileobj, filename, outpath, write.fun, ...)
}
\description{
Use write function to put a file object to a destination. Check automatically
if the destination is a remote url with the need to execute specific
functions.
}
