% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{read_file}
\alias{read_file}
\title{Load single fcs file into a file matrix}
\usage{
read_file(flow_entry, simple_marker_names = FALSE, dataset = 1)
}
\description{
Load single fcs file into a file matrix
}
