% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{read_files}
\alias{read_files}
\title{Load fcs files into a file matrix}
\usage{
read_files(file_matrix)
}
\arguments{
\item{file_matrix}{Matrix containing filename, group, id and set information as generated by get_dir().}
}
\value{
File matrix with loaded fcs files.
}
\description{
Load fcs files into a file matrix
}
\examples{
file_matrix = get_dir('/data', 'fcs')
read_files(file_matrix)
}
