% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{ReadDataset}
\alias{ReadDataset}
\title{Read single dataset}
\usage{
ReadDataset(path, remove.duplicates = T, filters = list(), ...)
}
\arguments{
\item{path}{Top directory which contains the lower directories serving as different groups.}

\item{remove.duplicates}{Remove duplicate file entries.}

\item{filters}{Optional filters used on slot of files.}
}
\value{
File matrix with specified set and no duplicates.
}
\description{
Convenient wapper to extract a single set without duplicates from directory file structure.
}
