% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{ReadDatasetJson}
\alias{ReadDatasetJson}
\title{Read jSON info file with additional case metainformation}
\usage{
ReadDatasetJson(json.path, dir.path = "", temp.path = "")
}
\description{
Create list:groups->list:case->S4:flowData
}
