% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{ReadDatasets}
\alias{ReadDatasets}
\title{Read all datasets}
\usage{
ReadDatasets(path, ...)
}
\arguments{
\item{path}{Dataset root directory.}
}
\description{
Data used by flowProc has to be organized into a specific hierarchy.
 BASE_DIR - Root directory of the data
 DATASET - Set of data containing different cohorts
 COHORTS - folder name is used to denote cohort affiliation
 FILES - filename used to infer label and tube_set
}
