% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicates from file matrix.}
\usage{
remove_duplicates(file_info)
}
\arguments{
\item{File}{matrix as output by get_dir()}
}
\value{
File matrix with duplicates removed. All occurrences are removed.
}
\description{
Remove duplicates from file matrix.
}
