% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{remove_marginal}
\alias{remove_marginal}
\title{Remove marginal cells}
\usage{
remove_marginal(flow_entry)
}
\arguments{
\item{flow_entry}{Flow data with loaded fcs slot.}
}
\value{
Flow entry with marginal events removed.
}
\description{
Removes all events with at least one border value.
The border is defined via the parameter range or as the min/max of the dataset.
}
\examples{
remove_marginal(testflow)
}
