% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_processing.R
\name{RemoveMarginal}
\alias{RemoveMarginal}
\title{Remove marginal cells}
\usage{
RemoveMarginal(flow_entry, lower = TRUE, upper = TRUE, ...)
}
\arguments{
\item{flow_entry}{Flow data with loaded fcs slot.}

\item{lower}{Boolean to enable lower boundary filtering.}

\item{upper}{Boolean to enable upper boundary filtering.}
}
\value{
Flow entry with marginal events removed.
}
\description{
Removes all events with at least one border value.
The border is defined via the parameter range or as the min/max of the dataset.
}
\examples{
RemoveMarginal(testflow)
}
