% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RowToS4}
\alias{RowToS4}
\title{Convert a single case row to a list of S4 objects for this case. Each
Filepath will be converted to a separate object.}
\usage{
RowToS4(case.row, dir.path, temp.path)
}
\description{
Convert a single case row to a list of S4 objects for this case. Each
Filepath will be converted to a separate object.
}
