% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{select_set}
\alias{select_set}
\title{Select one set in file matrix}
\usage{
select_set(file_info, set)
}
\arguments{
\item{file_info}{File Matrix as loaded by get_dir}

\item{set}{Specifies set which will be returned.}
}
\value{
File matrix containing specified set.
}
\description{
Select one set in file matrix
}
