% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{transform_ff}
\alias{transform_ff}
\title{Transform flowcytometric marker channels.}
\usage{
transform_ff(ff, sel_fun, trans_fun)
}
\arguments{
\item{ff}{Flowframe}

\item{sel_fun}{Selection function to determine transform application, which returns for an input name either TRUE or FALSE}

\item{trans_fun}{Transformation function, which will be applied.}
}
\value{
File Matrix with transformed flow frames.
}
\description{
Apply a transformation function to the channels specified by the selection function.
}
