% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{transform_ffs}
\alias{transform_ffs}
\title{Transform flowcytometric marker channels.}
\usage{
transform_ffs(ffs, sel_fun, trans_fun, threads = 1)
}
\arguments{
\item{ffs}{File matrix with loaded flow frames.}

\item{sel_fun}{Selection function to determine transform application, which returns for an input name either TRUE or FALSE}

\item{trans_fun}{Transformation function, which will be applied.}

\item{threads}{Number of cpu threads.}
}
\value{
File Matrix with transformed flow frames.
}
\description{
Apply a transformation function to the channels specified by the selection function.
}
