% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_processing.R
\name{TransformChannels}
\alias{TransformChannels}
\title{Transform flowcytometric marker channels.}
\usage{
TransformChannels(flowframe, selection.function, transform.function, ...)
}
\arguments{
\item{flowframe}{FCS Flowframe.}

\item{selection.function}{Selection function to determine transform application, which returns for an input name either TRUE or FALSE}

\item{transform.function}{Transformation function, which will be applied.}
}
\value{
File Matrix with transformed flow frames.
}
\description{
Apply a transformation function to the channels specified by the selection function.
}
