% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{create_file_info}
\alias{create_file_info}
\title{Read and preprocess file structure}
\usage{
create_file_info(path, ext, set)
}
\arguments{
\item{path}{Top directory which contains the lower directories serving as different groups.}

\item{ext}{File extension of files to be included. Generally this will be either lmd or fcs (case sensitive)}

\item{set}{Specifies set which will be returned.}
}
\value{
File matrix with specified set and no duplicates.
}
\description{
Convinient wapper to extract a single set without duplicates from directory file structure.
}
