% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{filter_flowFrame_majority}
\alias{filter_flowFrame_majority}
\title{Remove rare flowframe channels and exclude flowframes without common channels.}
\usage{
filter_flowFrame_majority(fcs_info, threshold)
}
\arguments{
\item{fcs_info}{File matrix with loaded flowframes.}

\item{threshold}{Minimum ratio for marker to be common.}
}
\value{
File matrix with removed channels.
}
\description{
Remove rare flowframe channels and exclude flowframes without common channels.
}
