% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{GroupBy}
\alias{GroupBy}
\title{Group list by specific field}
\usage{
GroupBy(upsampled.list, group.on, acc = slot, ...)
}
\arguments{
\item{upsampled.list}{List of entries.}

\item{group.on}{Slot to group entries on if using the default indexer and
comparator.}

\item{acc}{Alternative function to access the information for grouping.}
}
\description{
Group list by specific field
}
\examples{
GroupBy(all.files, 'label')
GroupBy(all.files, 'label', thread.num=12)
}
