% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{marker_occurrences}
\alias{marker_occurrences}
\title{Get distribution of marker channels across whole dataset.}
\usage{
marker_occurrences(fcs_info)
}
\arguments{
\item{fcs_info}{File matrix as loaded by get_dir().}
}
\value{
Table with occurrences of marker channels in histogram data.
}
\description{
Get distribution of marker channels across whole dataset.
}
