% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{process_single}
\alias{process_single}
\title{Remove debris
Load and filter single file info entry}
\usage{
process_single(file_entry, selection, simple_marker_names = FALSE,
  trans = "logicle", remove_margins = TRUE, upper = TRUE, lower = TRUE)
}
\arguments{
\item{selection}{Marker names selected from flowframe.}

\item{trans}{String name for transformation function.}

\item{remove_margins}{Boolean for removal of events with 0 or maximum values.}

\item{file_row}{File info row from a file matrix.}
}
\value{
File row or NULL if marker name not in flowframe.
}
\description{
Remove debris
Load and filter single file info entry
}
\examples{
t = process_single(files[[1]],selection)
}
