% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{process_single}
\alias{process_single}
\title{Load and filter single file info entry}
\usage{
process_single(file_row, selection)
}
\arguments{
\item{file_row}{File info row from a file matrix.}

\item{selection}{Marker names selected from flowframe.}
}
\value{
File row or NULL if marker name not in flowframe.
}
\description{
Load and filter single file info entry
}
