% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{read_files}
\alias{read_files}
\title{Load fcs files into a file matrix}
\usage{
read_files(file_info, threads = 1, simple_marker_names)
}
\arguments{
\item{file_info}{Matrix or Vector containing filename, group, id and set information as generated by get_dir().}

\item{threads}{Number of cpu threads used in file reading.}

\item{simple_marker_names}{Use antibody name as marker name.}
}
\value{
File matrix or vector with loaded fcs files.
}
\description{
Load fcs files into a file matrix
}
\examples{
file_matrix = get_dir('/data', 'fcs')
read_files(file_matrix)
}
