% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{read_files}
\alias{read_files}
\title{Load fcs files into a file matrix}
\usage{
read_files(file_entries, cluster, simple_marker_names = FALSE, dataset = 1)
}
\arguments{
\item{cluster}{Define cluster for concurrent operations.}

\item{simple_marker_names}{Use antibody name as marker name.}

\item{file_info}{Matrix or Vector containing filename, group, id and set
information as generated by GetDir().}

\item{Dataset}{used in fcs files with multiple datasets.}
}
\value{
File entries list or file entry object with loaded fcs files.
}
\description{
Load fcs files into a file matrix
}
\examples{
file_entries = GetDir('/data', 'fcs')
read_files(file_matrix)
}
