% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_processing.R
\name{ReadFile}
\alias{ReadFile}
\title{Load single fcs file into a file matrix}
\usage{
ReadFile(flow.entry, simple.marker.names = FALSE, dataset = 1, ...)
}
\description{
Load single fcs file into a file matrix
}
