% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicates with same filename from file matrix.}
\usage{
remove_duplicates(fcs_entries)
}
\arguments{
\item{File}{matrix as output by GetDir()}
}
\value{
File matrix with duplicates removed. All occurrences are removed.
}
\description{
Remove duplicates with same filename from file matrix.
}
\examples{
t = remove_duplicates(files)
}
