% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_input.R
\name{remove_small_cohorts}
\alias{remove_small_cohorts}
\title{Remove groups in file matrix with fewer entries than the threshold.}
\usage{
remove_small_cohorts(file_entries, minsize)
}
\arguments{
\item{minsize}{Threshold group size. Groups with fewer entries will be removed from the returned matrix.}

\item{fcs_info}{File matrix as generated by GetDir().}
}
\value{
File matrix without groups below threshold.
}
\description{
Remove groups in file matrix with fewer entries than the threshold.
}
\examples{
file_info = GetDir('../data', 'fcs')
file_info = remove_small_cohorts(file_info, 20)
t = remove_small_cohorts(files, 1000)
}
